move Models to Model.swift
